requirements.txt
#Django==3.1.6
Django==3.2.5
faker
django-phonenumber-field[phonenumbers]
psycopg2-binary>=2.8
django-bootstrap4
gunicorn==20.0.4